* Move the `_setup_commands` into the base class. * Misc small cleanup.
